Move request utils to core/manager module
